requirements.txt
binance-connector==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
coloredlogs==15.0.1
humanfriendly==10.0
idna==3.7
numpy==2.0.1
pandas==2.2.2
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
ta==0.11.0
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
